backend/requirements.txt
Django==5.0
djangorestframework
channels==4.0
channels-redis
numpy
